sdk/ai/azure-ai-agentserver-agentframework/dev_requirements.txt
-e ../../../eng/tools/azure-sdk-tools
../../identity/azure-identity
../azure-ai-agentserver-core
python-dotenv
pywin32